Clean up projectile files

+++ sdl-game/src/vector2.h
@@ -15,10 +15,10 @@ public:
 	float dot_product(Vector2 a, Vector2 b);
 	Vector2 normalize(Vector2 a);
 
-	int inline getX() { return _x; }
-	int inline getY() { return _y; }
-	void inline setX(int x) { _x = x; }
-	void inline setY(int y) { _y = y; }
+	float inline getX() { return _x; }
+	float inline getY() { return _y; }
+	void inline setX(float x) { _x = x; }
+	void inline setY(float y) { _y = y; }
 
 private:
 	float _x, _y = 0;
